requirements.txt
alembic==1.4.2
aniso8601==8.0.0
attrs==19.3.0
bcrypt==3.1.7
cffi==1.14.0
click==7.1.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Migrate==2.5.3
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-Testing==0.8.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
Mako==1.1.2
MarkupSafe==1.1.1
pycparser==2.20
PyJWT==1.7.1
pyrsistent==0.16.0
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.1
six==1.14.0
SQLAlchemy==1.3.16
Werkzeug==0.16.1
